requirements.txt
torch>=2.1.0
torchaudio>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
peft>=0.10.0
accelerate>=0.28.0
sentencepiece>=0.2.0
protobuf>=3.20.0
soundfile>=0.12.0
librosa>=0.10.0
audioread>=3.0.0
pyyaml>=6.0
sacrebleu>=2.4.0
unbabel-comet>=2.2.0
evaluate>=0.4.0
jiwer>=3.0.0
tensorboard>=2.15.0
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.0.0
gitpython>=3.1.0
huggingface-hub>=0.21.0
scipy>=1.11.0
